Add files via upload
Login SQL is the MYSQLITE module to create and log in. KV file is for styling and .py is the meat of the program. Might need to edit but created the initial log in screen. NEWUSER is the module to create a NEW USER. Might need to modify the script a bit but got rid of most the references to my chat app. -Zach